Ensure that rebin returns a 16 bit image

+++ TiffSequence.py
@@ -565,7 +565,7 @@ def loadTimes(filename,firstFrameIndex=0,firstTimeValue=0,scaleFactor=1.0):
 
 def rebin(a, shape):
     sh = shape[0],a.shape[0]//shape[0],shape[1],a.shape[1]//shape[1]
-    return a.reshape(sh).sum(-1).sum(1)
+    return a.reshape(sh).sum(-1).sum(1).astype(np.uint16)
 
 class TimesDict(dict):
 	"""
